requirements.txt
socket
facilities
mysql.connector
threading
